Move repository test libraries into lib/tests and export as .tests.*
