+++ index.js
@@ -1,6 +1,5 @@
 'use strict';
 
 const jsTools = require('./lib/jsonschema-tools.js');
-const tests   = require('./lib/tests.js');
 
-module.exports = { ...jsTools, ...tests };
+module.exports = { ...jsTools, tests: require('./lib/tests.js') };
